Implementierung von verschiedenen ausführung für die get_Call Methode, außerdem wird die SVM als Grafik am ende gespeichert

+++ Mia/Control/primitiveki.cpp
@@ -50,7 +50,7 @@ int PrimitiveKI::getCall(Value last, Value rand){
         result = rand.toInt();
     }
     else{
-        result = getRandomValueGreaterX(last.toInt());
+        result = getRandomValueGreaterX(last);
     }
     return result;
 }
